Auto Start for windows added
Added auto start option for windows GUI. in the GUI settings added
"Start automatically with Windows" option, which creates an auto run key
in the windows registry for the current user. Tested on windows 8. This
uses the current running location of the GUI application so it does not
need to be installed to work. a check for mono runtime is performed and
it just hides the option if the GUI application is not running natively
in Windows.